# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Native-text PDF or DOCX",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post",
        "type": "object"
      },
...
      "DictionaryTermApprovalRequest": {
        "additionalProperties": false,
        "description": "Admin approval or rejection of a proposed term.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "$ref": "#/components/schemas/DictionaryTermStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "DictionaryTermApprovalRequest",
        "type": "object"
      },
      "DictionaryTermCreate": {
        "additionalProperties": false,
        "description": "Request to propose a term for a project or organization.",
        "properties": {
          "rationale": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "scope": {
            "pattern": "^(organization|project:[0-9a-fA-F-]{36})$",
            "title": "Scope",
            "type": "string"
          },
          "term": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "term",
          "scope"
        ],
        "title": "DictionaryTermCreate",
        "type": "object"
      },
      "DictionaryTermListResponse": {
        "additionalProperties": false,
        "description": "Paginated dictionary collection.",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/DictionaryTermRead"
            },
            "title": "Terms",
            "type": "array"
          }
        },
        "required": [
          "terms",
          "pagination"
        ],
        "title": "DictionaryTermListResponse",
        "type": "object"
      },
      "DictionaryTermRead": {
        "additionalProperties": false,
        "description": "Governed dictionary entry.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DictionaryTermStatus"
          },
          "term": {
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "id",
          "term",
          "scope",
          "status",
          "created_at"
        ],
        "title": "DictionaryTermRead",
        "type": "object"
      },
...
      "DocumentListResponse": {
        "additionalProperties": false,
        "description": "Paginated document collection.",
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentRead"
            },
            "title": "Documents",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "documents",
          "pagination"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
      "DocumentRead": {
        "additionalProperties": false,
        "description": "Safe document metadata returned to clients.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "progress_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Pct",
            "type": "integer"
          },
          "review_status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "media_type",
          "size_bytes",
          "sha256",
          "status",
          "review_status",
          "progress_pct",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentRead",
        "type": "object"
      },
...
      "DocumentStatusResponse": {
        "additionalProperties": false,
        "description": "Processing and stage status used for polling.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "progress_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Pct",
            "type": "integer"
          },
          "review_status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRunRead"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "review_status",
          "progress_pct",
          "stages",
          "updated_at"
        ],
        "title": "DocumentStatusResponse",
        "type": "object"
      },
      "DocumentUploadResponse": {
        "additionalProperties": false,
        "description": "Response after a valid upload is persisted and queued.",
        "examples": [
          {
            "created_at": "2026-09-04T04:00:00Z",
            "filename": "Inspection_Report_Rev-A.pdf",
            "id": "ed846165-102f-49f9-9fd0-f25c0d4cfe6e",
            "status": "QUEUED"
          }
        ],
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          }
        },
        "required": [
          "id",
          "filename",
          "status",
          "created_at"
        ],
        "title": "DocumentUploadResponse",
        "type": "object"
      },
      "FieldViolation": {
        "additionalProperties": false,
        "description": "One request-validation failure safe to expose to an API client.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "code"
        ],
        "title": "FieldViolation",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IssueCategory": {
        "description": "Top-level issue grouping used by API filters and the review UI.",
        "enum": [
          "LINGUISTIC",
          "TRACEABILITY",
          "SYSTEM"
        ],
        "title": "IssueCategory",
        "type": "string"
      },
...
      "IssueListResponse": {
        "additionalProperties": false,
        "description": "Paginated issue collection plus review summary counts.",
        "properties": {
          "counts_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Severity"
            },
            "title": "Counts By Severity",
            "type": "object"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueRead"
            },
            "title": "Issues",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "issues",
          "pagination",
          "counts_by_severity"
        ],
        "title": "IssueListResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "RFC 9457-style error response with a stable application code.",
        "examples": [
          {
            "code": "FEATURE_NOT_READY",
            "detail": "Document upload is not implemented yet.",
            "errors": [],
            "instance": "/api/v1/documents/upload",
            "status": 501,
            "title": "Feature not ready",
            "type": "urn:docqc:error:FEATURE_NOT_READY"
          }
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "TraceabilitySummaryResponse": {
        "additionalProperties": false,
        "description": "Counts required by the audit-focused dashboard.",
        "properties": {
          "counts_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts By Severity",
            "type": "object"
          },
          "counts_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts By Type",
            "type": "object"
          },
          "critical_count": {
            "minimum": 0.0,
            "title": "Critical Count",
            "type": "integer"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "unresolved_count": {
            "minimum": 0.0,
            "title": "Unresolved Count",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "counts_by_type",
          "counts_by_severity",
          "critical_count",
          "unresolved_count"
        ],
        "title": "TraceabilitySummaryResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Automated first-pass quality review of engineering documents (PDF/DOCX)",
    "title": "Document QC API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/dictionary/terms": {
      "get": {
        "description": "List governed dictionary terms with optional scope filtering.",
        "operationId": "list_dictionary_terms_api_v1_dictionary_terms_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryTermListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "List Dictionary Terms",
        "tags": [
          "dictionary"
        ]
      },
      "post": {
        "description": "Propose a project- or organization-scoped dictionary term.",
        "operationId": "create_dictionary_term_api_v1_dictionary_terms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryTermCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryTermRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Create Dictionary Term",
        "tags": [
          "dictionary"
        ]
      }
    },
    "/api/v1/dictionary/terms/{term_id}/approve": {
      "patch": {
        "description": "Approve or reject a proposed dictionary term.",
        "operationId": "approve_dictionary_term_api_v1_dictionary_terms__term_id__approve_patch",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryTermApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryTermRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Approve Dictionary Term",
        "tags": [
          "dictionary"
        ]
      }
    },
    "/api/v1/documents": {
      "get": {
        "description": "List documents visible to the current user.",
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "description": "Persist a validated document and enqueue canonical extraction.",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "description": "Return safe metadata for one document.",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/export": {
      "get": {
        "description": "Export the annotated rendition or structured issue log.",
        "operationId": "export_document_api_v1_documents__document_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": true,
            "schema": {
              "pattern": "^(pdf|xlsx|csv)$",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Export Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/issues": {
      "get": {
        "description": "List issues with optional canonical category filtering.",
        "operationId": "list_document_issues_api_v1_documents__document_id__issues_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "List Document Issues",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/status": {
      "get": {
        "description": "Return processing progress and latest stage states.",
        "operationId": "get_document_status_api_v1_documents__document_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Get Document Status",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/traceability-summary": {
      "get": {
        "description": "Return traceability counts for audit triage.",
        "operationId": "get_traceability_summary_api_v1_documents__document_id__traceability_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceabilitySummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Feature not implemented"
          }
        },
        "summary": "Get Traceability Summary",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}